+++ src/env-var/cc-env-var-linked-services.js
@@ -122,7 +122,7 @@ export class CcEnvVarLinkedServices extends LitElement {
         </div>
       ` : ''}
       
-      ${this.services != null && !this.error && this.services.length > 0 ? html`
+      ${!this.error && this.services?.length > 0 ? html`
         <div class="service-list">
           ${this.services.map((s) => html`
             <cc-env-var-form readonly .variables=${s.variables} heading=${this._getServiceHeading(s.name)} error="${ifDefined(s.error)}">